need to test:
This is how i've written the firmware to receive the value for raceLengthTicks from OpenSprints:

the micro waits for the 'l' character. that's the lower case letter that begins the word 'letter', not a 1 or an upper case i.
then the next two characters it receives need to contain the value for raceLengthTicks.
* the first character is raceLengthTicks modulo 256.
* the second character is Floor(raceLengthTicks/256).
then it looks for the next character to be carriage return, or '\r'.
when it receives all that, it sends back the string "OK 12345\r" where 12345 would be raceLengthTicks back in decimal form.

we probably don't need to end the two numerical chars with the carriage return. the maximum value of the byte being sent representing raceLengthTicks would correspond to a 10 mile race for rollers with 0.25 meter circumference.

I'm thinking that this interaction happens when a user presses the "Start Race" button. OpenSprints sends the 4 characters for the raceTickLength message. OpenSprints waits to receive the "OK 12345\r" message before sending the micro the 'g' command to start the race.
It seems safest to set the raceLengthTicks right before every race, although that step could be taken at other idle moments.